Squashed commit of the following:
commit 641958d084562c2e9087186b647449e0d715132c

    Rearrange fns to make merge back cleaner.

commit 801f8e4cc989877c39a92a8b55a1021ab8ad5366

    Fix benign warning.

commit 8d2a092198c2146168ccb7981ed82e0d101ea971

    Clean up hooking.

    Detours is now used only for add_jmp; add_iat is back to using hook_iat.
    This also removes the need for detach(), since undoing a hook_iat is
    simple.

commit 7e30cfbc781908a6a77ab428b1a75b8405c9cb85

    Repair our IAT after hooking.

    Clink is working now, using Detours.

commit 71186e4ae18aea0cd99c1347431dbf324da53554

    Use Detours to hook APIs.

    This compiles and runs, but isn't hooking kernelbase!ReadConsoleW quite
    right yet and so it isn't intercepting CMD's calls yet.  And it isn't
    repairing the IAT correctly yet.

+++ readline/compat/hooks.c
@@ -27,7 +27,7 @@ static int mb_to_wide(const char* mb, wchar_t* fixed_wide, size_t fixed_size, wc
 
     wchar_t* wide;
     int alloced;
-    if (needed <= fixed_size)
+    if ((size_t)needed <= fixed_size)
     {
         wide = fixed_wide;
         alloced = 0;
